Fix transfer to stream HLS test
